[CC] documentation and function renaming
