+++ validations/loudness_ecma/valid_ear_filter_design.py
@@ -11,7 +11,7 @@ from numpy import (
 import matplotlib.pyplot as plt
 
 from mosqito.functions.loudness_ecma_spain.ear_filter_design import ear_filter_design
-from mosqito.functions.loudness_ecma_spain.sine_wave_generator import (
+from mosqito.functions.shared.sine_wave_generator import (
     sine_wave_generator,
 )
